move down encoding stuff
