+++ src/lib/base58-encodings.ts
@@ -1,0 +1,7 @@
+import { fieldEncodings } from './base58.js';
+import { Field } from '../snarky.js';
+
+export { TokenId, ReceiptChainHash, LedgerHash, EpochSeed, StateHash };
+
+const { TokenId, ReceiptChainHash, EpochSeed, LedgerHash, StateHash } =
+  fieldEncodings(Field);
