+++ src/lib/mina/account.ts
@@ -3,7 +3,7 @@ import { Bool, Field } from '../../snarky.js';
 import { Permissions } from '../account_update.js';
 import { UInt32, UInt64 } from '../int.js';
 import { PublicKey } from '../signature.js';
-import * as Encoding from '../encoding.js';
+import { TokenId, ReceiptChainHash } from '../base58-encodings.js';
 import { genericLayoutFold } from '../../snarkyjs-bindings/provable/from-layout.js';
 import {
   customTypes,
@@ -134,13 +134,12 @@ function parseFetchedAccount({
     provedState;
   let partialAccount: PartialAccount = {
     publicKey: PublicKey.fromBase58(publicKey),
-    tokenId: Encoding.TokenId.fromBase58(token),
+    tokenId: TokenId.fromBase58(token),
     tokenSymbol: tokenSymbol ?? undefined,
     balance: balance && UInt64.from(balance.total),
     nonce: UInt32.from(nonce),
     receiptChainHash:
-      (receiptChainHash &&
-        Encoding.ReceiptChainHash.fromBase58(receiptChainHash)) ||
+      (receiptChainHash && ReceiptChainHash.fromBase58(receiptChainHash)) ||
       undefined,
     delegate:
       (delegateAccount && PublicKey.fromBase58(delegateAccount.publicKey)) ??
